fix: change default text color to black
I'd rather have it be RESET, as it's more accurate. However, it looks like there's a bug with wrapping lines and formatting codes, where RESET only resets modifiers to what they were at the start of the line.

As I want the resulting books to be compatible with Vanilla Minecraft, I can't fix this. So we'll have to work around it.